test: add coverage for metadata redaction
