+++ bugsnag-android-core/src/test/resources/metadata_redaction_0.json
@@ -1,0 +1,7 @@
+{
+  "example": {
+    "normal": "safe",
+    "password": "[REDACTED]",
+    "confirm_password": "[REDACTED]"
+  }
+}
