+++ bugsnag-android-core/src/test/resources/metadata_redaction_1.json
@@ -1,0 +1,9 @@
+{
+  "example": {
+    "sensitiveMap": {
+      "normal": "safe",
+      "password": "[REDACTED]",
+      "confirm_password": "[REDACTED]"
+    }
+  }
+}
