+++ bugsnag-android-core/src/test/resources/metadata_redaction_2.json
@@ -1,0 +1,5 @@
+{
+  "foo": {
+    "password": "[REDACTED]"
+  }
+}
