+++ bugsnag-android-core/src/test/resources/metadata_redaction_3.json
@@ -1,0 +1,1 @@
+{}
