+++ bugsnag-android-core/src/test/resources/metadata_redaction_4.json
@@ -1,0 +1,6 @@
+{
+  "foo": {
+    "bar": "[REDACTED]",
+    "password": "abc123"
+  }
+}
